requirements.txt
certifi~=2022.12.7
Flask~=2.2.2
pymongo~=4.3.3
gunicorn
python-dotenv~=0.21.0